fix: refactoring to meet requested changes
