+++ bundles/com.espressif.idf.debug.gdbjtag.openocd/plugin.xml
@@ -166,5 +166,14 @@
           label="IDF Process Console">
     </consoleFactory>
  </extension>
+ <extension
+       point="org.eclipse.debug.core.statusHandlers">
+    <statusHandler
+          class="com.espressif.idf.debug.gdbjtag.openocd.ui.OpenocdStatusHandler"
+          code="5012"
+          id="com.espressif.idf.debug.gdbjtag.openocd.openocdStatusHandler"
+          plugin="com.espressif.idf.debug.gdbjtag.openocd">
+    </statusHandler>
+ </extension>
 
 </plugin>
